feat: add list of available profiles to run error
